backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
openai==1.3.0
streamlit==1.28.1
unstructured==0.10.30
pypdf==3.17.0
tiktoken==0.5.1